+++ Tests/Tests/STPCardTest.m
@@ -283,6 +283,7 @@
     XCTAssertEqual(card.expYear, (NSUInteger)2017);
     XCTAssertEqual(card.funding, STPCardFundingTypeCredit);
     XCTAssertEqualObjects(card.last4, @"4242");
+    XCTAssertEqualObjects(card.metadata, @{@"order_id": @"6735"});
     XCTAssertEqualObjects(card.name, @"Jane Austen");
 
     XCTAssertNotEqual(card.allResponseFields, response);